db schema relationship update
Updated the schema so that a user can have multiple devices, few plans can be assigned to many phones, but only one line per phone. a single device will have one plan and one line that can have null values.  run the db and see if this is what we want.